requirements.txt
py-clob-client==0.19.0
python-telegram-bot==20.7
aiohttp==3.9.1
aiohttp-cors==0.7.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
colorlog==6.8.0
web3==6.13.0
eth-account==0.10.0
cachetools==5.3.2
tenacity==8.2.3
